+++ store-service/src/main/java/practice/store/product/ProductService.java
@@ -2,21 +2,19 @@ package practice.store.product;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.PropertySource;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import practice.store.order.details.OrderProductPayload;
 import practice.store.utils.converter.EntitiesConverter;
 import practice.store.utils.converter.PayloadsConverter;
 import practice.store.utils.numbers.CalculatePrice;
+import practice.store.validator.ProductValidator;
 
 import javax.persistence.EntityNotFoundException;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
-@PropertySource("classpath:productsDiscountValue.properties")
 @RequiredArgsConstructor
 @Transactional
 @Service
@@ -32,10 +30,7 @@ public class ProductService {
 
     private final AvailabilityService availabilityService;
 
-    @Value("${discount.percentage.max.value.higher}")
-    private int DISCOUNT_PERCENTAGE_MAX;
-    @Value("${discount.percentage.max.value.lower}")
-    private int DISCOUNT_PERCENTAGE_MIN;
+    private final ProductValidator productValidator;
 
 
     public ProductPayload getByUuid(String uuid) {
@@ -59,13 +54,13 @@ public class ProductService {
     }
 
     public boolean save(ProductPayload productPayload) {
-        if (isProductUuidExist(productPayload.getProductUUID())) {
+        if (productValidator.isProductUuidExist(productPayload.getProductUUID())) {
             return false;
         }
-        if (!isProductWithdrawFromSale(productPayload.getAvailability())) {
+        if (!productValidator.isProductWithdrawFromSale(productPayload.getAvailability())) {
             return false;
         }
-        if (!areDiscountAndPriceCorrect(productPayload)) {
+        if (!productValidator.areDiscountAndPriceCorrect(productPayload)) {
             return false;
         }
 
@@ -87,7 +82,7 @@ public class ProductService {
         ProductEntity existingProduct = payloadsConverter.convertProduct(productPayload, calculatedAvailability, existProductEntityID);
 
         if (existingProduct.isHasDiscount()) {
-            if (!isDiscountPercentageInRangeCorrect(existingProduct.getDiscountPercentage())) {
+            if (!productValidator.isDiscountPercentageInRangeCorrect(existingProduct.getDiscountPercentage())) {
                 return false;
             }
             setFinalPrice(existingProduct);
@@ -109,16 +104,6 @@ public class ProductService {
     }
 
 
-    private boolean areDiscountAndPriceCorrect(ProductPayload productPayload) {
-        if (productPayload.isHasDiscount()) {
-            return isDiscountPercentageInRangeCorrect(productPayload.getDiscountPercentage())
-                    && isFinalPriceCorrect(productPayload.getBasePrice(), productPayload.getDiscountPercentage(), productPayload.getFinalPrice());
-        } else {
-            return isDiscountPercentageInRange(productPayload.getDiscountPercentage())
-                    && areFinalPriceAndBasePriceEquals(productPayload.getFinalPrice(), productPayload.getBasePrice());
-        }
-    }
-
     private void setFinalPrice(ProductEntity productEntity) {
         BigDecimal finalPrice = calculateFinalPrice.calculateFinalPrice(productEntity.getBasePrice(), productEntity.getDiscountPercentage());
         productEntity.setFinalPrice(finalPrice);
@@ -127,44 +112,5 @@ public class ProductService {
     private void setNoDiscount(ProductEntity productEntity) {
         productEntity.setDiscountPercentage(0);
         productEntity.setFinalPrice(productEntity.getBasePrice());
-    }
-
-    private boolean areFinalPriceAndBasePriceEquals(BigDecimal finalPrice, BigDecimal basePrice) {
-        boolean result = finalPrice.compareTo(basePrice) == 0;
-        return logIfFalse(result, String.format("Final price: %.2f and base price: %.2f are not equal", finalPrice, basePrice));
-    }
-
-    private boolean isDiscountPercentageInRangeCorrect(int discountPercentage) {
-        boolean result = (discountPercentage <= DISCOUNT_PERCENTAGE_MAX) && (discountPercentage >= DISCOUNT_PERCENTAGE_MIN);
-        return logIfFalse(result, String.format("Discount percentage is incorrect. Is: %d", discountPercentage));
-    }
-
-    private boolean isDiscountPercentageInRange(int discountPercentage) {
-        boolean result = discountPercentage == 0;
-        return logIfFalse(result, String.format("Discount percentage is not zero. Is: %d", discountPercentage));
-    }
-
-    private boolean isFinalPriceCorrect(BigDecimal basePrice, int discountPercentage, BigDecimal finalPrice) {
-        BigDecimal finalPriceCalculate = calculateFinalPrice.calculateFinalPrice(basePrice, discountPercentage);
-        boolean result = finalPriceCalculate.compareTo(finalPrice) == 0;
-        return logIfFalse(result, String.format("Final price is incorrect. Is: %.2f but should be: %.2f", finalPrice, finalPriceCalculate));
-    }
-
-    private boolean isProductUuidExist(String uuid) {
-        boolean result = productRepository.existsByProductUUID(uuid);
-        return logIfFalse(result, String.format("Product with UUID: %s is not exist", uuid));
-    }
-
-    private boolean isProductWithdrawFromSale(Availability availability) {
-        boolean result = !(availability.equals(Availability.WITHDRAW_FROM_SALE));
-        return logIfFalse(result, "Product is withdraw from sale");
-    }
-
-    private boolean logIfFalse(boolean result, String desc) {
-        if (!result) {
-            log.error("Result condition is false because: {}.", desc);
-            return false;
-        }
-        return true;
     }
 }
