+++ store-service/src/main/java/practice/store/validator/BaseValidator.java
@@ -1,0 +1,6 @@
+package practice.store.validator;
+
+abstract class BaseValidator {
+
+    abstract boolean logIfFalse(boolean result, String desc);
+}
